feat: Replace styled-components with emotion (#230)
* Replace styled-components with emotion

* Upgrade styled-system

* Fix calendar & error message

* Fix nested styled getters

* Fix broken tests

* Update snapshots

* Add jest-emotion

* Add emotion to peerDependencies

* Fix ci test scripts

Tests scripts aren't in sub packages so we don't want to be running them with lerna

* Replace styledOmitProps with omitProps

Why? Wrapping emotion styled is not compatible with emotions babel plugin

* Run build before lint to fix import linting errors

This is a temp fix. The link command should be run in the previously buit container

* Build app in dockerfile

* Run lerna bootstrap in Dockerfile

* Disable import eslint rules

Import rules are failing in CI, likely related to a issue with lerna bootstrap.

We should re-enable these rules when we move to a mono package