remove internal state from ButtonGroup (#346)
ButtonGroup was managing its value/checked state internally while also receiving the value as a prop. This meant that any changes to value would not be reflected in the checked state of the ButtonGroup after the initial render. This change removes the internal state and relies solely on the value prop.