refactor: use defaultProps in Flex component
We are currently hard coding the display value to `flex` in the `Flex` component. This is what is required 95% of the time, but it limits the ability to show and hide the component across breakpoints. This change to using defaultProps rather than hard coded properties will make it possible to hide a flexbox on the lowest break point and show it on all others. As well cater to other rare, but useful, requirements.